Windows port: op-family MSVC fixes (narrowing casts, builtins, force-inline)
- explicit casts where MSVC hard-errors on narrowing in braced initializers
  (GNU/Clang builds run these files with -Wno-narrowing): data_movement
  (concat, copy, gather, indexed_fill, moe remaps, roll, scatter, sort,
  transpose), eltwise (binary_ng, ternary, unary), matmul factories,
  reduction (accumulation, topk), core work_split
- __builtin_unreachable() call sites routed through ttsl::unreachable()
- __builtin_clz replaced with std::bit_floor in topk_utils
- bare glibc 'uint' spelled uint32_t in unary_op_utils
- slice_device_operation uses TT_FORCE_INLINE instead of raw
  __attribute__((always_inline))

(cherry picked from commit 605eba2eeed5ecad58cbdd581f2437522d2d79c3)

+++ ttnn/cpp/ttnn/operations/data_movement/gather/tosa/gather_tosa.cpp
@@ -24,7 +24,10 @@ Tensor pre_tosa_gather_transform_input_index_tensor(const Tensor& input_tensor, 
     // Unsqueeze the input tensor to add a new dimension
     const Tensor unsqueezed_tensor = ttnn::unsqueeze(input_tensor, -1);
     // Create a shape vector for the new tensor
-    ttsl::SmallVector<int32_t> shape_vector = {input_tensor.logical_shape()[0], input_tensor.logical_shape()[1], C};
+    ttsl::SmallVector<int32_t> shape_vector = {
+        static_cast<int32_t>(input_tensor.logical_shape()[0]),
+        static_cast<int32_t>(input_tensor.logical_shape()[1]),
+        static_cast<int32_t>(C)};
     Tensor expanded_tensor = ttnn::expand(unsqueezed_tensor, shape_vector, unsqueezed_tensor.memory_config());
 
     return expanded_tensor;
